dataflow/requirements.txt
apache-beam==2.40.0
google-api-core==1.31.6
google-api-python-client==2.51.0
google-apitools==0.5.31
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-appengine-logging==1.1.2
google-cloud-audit-log==0.2.2
google-cloud-bigquery
google-cloud-bigquery-storage==2.13.2
google-cloud-bigtable==1.7.2
google-cloud-core
google-cloud-language==1.3.2
google-cloud-logging==3.1.2
google-cloud-storage
google-cloud-translate==3.7.4
google-cloud-videointelligence==1.16.3
google-cloud-vision==1.0.2
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
greenlet==1.1.2
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-gcp==0.2.2
grpcio-status==1.47.0
h5py==3.7.0
hdfs==2.7.0
Werkzeug==2.1.2
wrapt==1.14.1
yapf==0.32.0
yarl==1.7.2
zipp==3.8.0
